# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Alpha Lens Investment Intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recommendations/events:
    get:
      operationId: getGlobalEvents
      tags: [recommendations]
      summary: Get recent global market events
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: List of global events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventsResponse"
  /recommendations/watchlist:
    get:
      operationId: getWatchlist
      tags: [recommendations]
      summary: Get watchlist items
      responses:
        "200":
          description: Watchlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistResponse"
    post:
      operationId: addToWatchlist
      tags: [recommendations]
      summary: Add asset to watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistAddRequest"
      responses:
        "200":
          description: Added to watchlist
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /recommendations/resolve-outcomes:
    post:
      operationId: resolveOutcomes
      tags: [recommendations]
      summary: Trigger automated outcome resolution run (admin only)
      responses:
        "200":
          description: Resolution digest
          content:
            application/json:
              schema:
                type: object
                properties:
                  digest:
                    $ref: "#/components/schemas/ResolutionDigest"
                required: [digest]
        "401":
          description: Authentication required
        "403":
          description: Admin access required
  /recommendations/{id}/outcome:
    patch:
      operationId: updateRecommendationOutcome
      tags: [recommendations]
      summary: Update outcome on a recommendation (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutcomeUpdateRequest"
      responses:
        "200":
          description: Updated recommendation
          content:
            application/json:
              schema:
                type: object
                properties:
                  recommendation:
                    $ref: "#/components/schemas/Recommendation"
                required:
                  - recommendation
        "401":
          description: Authentication required
        "403":
          description: Admin access required
        "404":
          description: Recommendation not found
  /leaderboard:
    get:
      operationId: getLeaderboard
      tags: [leaderboard]
      summary: Public signal accuracy leaderboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: type
          in: query
          schema:
            type: string
            enum: [trade, watch, avoid, all]
            default: all
        - name: status
          in: query
          schema:
            type: string
            enum: [resolved, open, all]
            default: all
      responses:
        "200":
          description: Leaderboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"
  /recommendations/watchlist/{id}:
    delete:
      operationId: removeFromWatchlist
      tags: [recommendations]
      summary: Remove asset from watchlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed from watchlist
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
...
  /trading/route/{recommendationId}:
    get:
      operationId: getRoutingDecision
      tags: [trading]
      summary: Preview which platform a recommendation routes to
      parameters:
        - name: recommendationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Routing decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutingDecisionResponse"
...
components:
  schemas:
...
    Recommendation:
      type: object
      properties:
        id:
          type: integer
        briefingId:
          type: integer
          nullable: true
        type:
          type: string
          enum: [trade, watch, avoid]
        urgency:
          type: string
          enum: [high, medium, low]
        title:
          type: string
        assetId:
          type: integer
          nullable: true
        assetTitle:
          type: string
        assetClass:
          type: string
        sector:
          type: string
        region:
          type: string
        direction:
          type: string
        aiProbability:
          type: number
          nullable: true
        marketPrice:
          type: number
          nullable: true
        assetPriceAtCall:
          type: number
          nullable: true
        edge:
          type: number
          nullable: true
        edgeType:
          type: string
          enum: [probability_gap, directional_conviction]
          nullable: true
        convictionScore:
          type: number
          nullable: true
        edgeCalculatedAt:
          type: string
          format: date-time
          nullable: true
        edgeAgeMinutes:
          type: integer
          nullable: true
        headline:
          type: string
        why:
          type: array
          items:
            type: string
        historicalContext:
          type: string
        bearCase:
          type: string
        entryTrigger:
          type: string
        confidence:
          type: integer
        window:
          type: string
        urgencyReason:
          type: string
        outcome:
          type: string
          enum: [correct, incorrect, partial]
          nullable: true
        resolutionDate:
          type: string
          format: date-time
          nullable: true
        resolutionNote:
          type: string
          nullable: true
        marketPriceAtResolution:
          type: number
          nullable: true
        paperReturn:
          type: number
          nullable: true
        resolutionMethod:
          type: string
          enum: [auto, manual]
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - type
        - title
        - confidence
    GlobalEvent:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        region:
          type: string
        impactLevel:
          type: string
          enum: [critical, high, medium, low]
        detail:
          type: string
        affectedAssets:
          type: array
          items:
            type: string
        direction:
          type: string
          enum: [bullish, bearish, mixed]
        timeContext:
          type: string
        scannedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - title
...
    EventsResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/GlobalEvent"
      required:
        - events
    WatchlistItem:
      type: object
      properties:
        id:
          type: integer
        assetId:
          type: integer
          nullable: true
        assetTitle:
          type: string
        assetClass:
          type: string
        alertEdgeThreshold:
          type: number
        notes:
          type: string
        addedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
    WatchlistResponse:
      type: object
      properties:
        watchlist:
          type: array
          items:
            $ref: "#/components/schemas/WatchlistItem"
      required:
        - watchlist
    WatchlistAddRequest:
      type: object
      properties:
        assetId:
          type: integer
        assetTitle:
          type: string
        assetClass:
          type: string
        alertEdgeThreshold:
          type: number
        notes:
          type: string
      required:
        - assetId
...
    RoutingDecisionResponse:
      type: object
      properties:
        recommendationId:
          type: integer
        title:
          type: string
        selectedPlatform:
          type: string
        reason:
          type: string
        tradeable:
          type: boolean
        usJurisdictionMode:
          type: boolean
        requireApproval:
          type: boolean
...
    OutcomeUpdateRequest:
      type: object
      properties:
        outcome:
          type: string
          enum: [correct, incorrect, partial]
        resolutionDate:
          type: string
          format: date-time
        resolutionNote:
          type: string
        marketPriceAtResolution:
          type: number
          nullable: true
        paperReturn:
          type: number
          nullable: true
      required:
        - outcome
        - resolutionDate
        - resolutionNote
    LeaderboardStats:
      type: object
      properties:
        trackRecordStart:
          type: string
        trackRecordEnd:
          type: string
        daysElapsed:
          type: integer
        daysRemaining:
          type: integer
        totalCalls:
          type: integer
        resolvedCalls:
          type: integer
        openCalls:
          type: integer
        correctCalls:
          type: integer
        incorrectCalls:
          type: integer
        partialCalls:
          type: integer
        winRate:
          type: number
        winRateWithPartial:
          type: number
        avgEdge:
          type: number
        avgAiProbability:
          type: number
        avgConvictionScore:
          type: number
        highConvictionWinRate:
          type: number
          nullable: true
        lowConvictionWinRate:
          type: number
          nullable: true
        totalPaperReturn:
          type: number
        paperReturnPct:
          type: number
        highConfidenceWinRate:
          type: number
          nullable: true
        highEdgeWinRate:
          type: number
          nullable: true
        paperReturnReliability:
          type: string
          enum: [verified, estimated, unavailable]
        autoResolved:
          type: integer
        manualResolved:
          type: integer
        pendingResolution:
          type: integer
        byType:
          type: object
          properties:
            trade:
              type: object
              properties:
                total:
                  type: integer
                resolved:
                  type: integer
                correct:
                  type: integer
                winRate:
                  type: number
              required: [total, resolved, correct, winRate]
            watch:
              type: object
              properties:
                total:
                  type: integer
                resolved:
                  type: integer
              required: [total, resolved]
            avoid:
              type: object
              properties:
                total:
                  type: integer
                resolved:
                  type: integer
                correct:
                  type: integer
                winRate:
                  type: number
              required: [total, resolved, correct, winRate]
          required: [trade, watch, avoid]
      required:
        - trackRecordStart
        - trackRecordEnd
        - daysElapsed
        - daysRemaining
        - totalCalls
        - resolvedCalls
        - openCalls
        - correctCalls
        - incorrectCalls
        - partialCalls
        - winRate
        - winRateWithPartial
        - avgEdge
        - avgAiProbability
        - totalPaperReturn
        - paperReturnPct
        - paperReturnReliability
        - byType
    LeaderboardCalibrationBucket:
      type: object
      properties:
        bucket:
          type: string
        calls:
          type: integer
        correct:
          type: integer
        rate:
          type: number
      required: [bucket, calls, correct, rate]
    ResolutionDigestEntry:
      type: object
      properties:
        id:
          type: integer
        assetTitle:
          type: string
        outcome:
          type: string
          enum: [correct, incorrect, partial]
        platform:
          type: string
          enum: [kalshi, polymarket, price, macro, unknown]
      required: [id, assetTitle, outcome, platform]
    ResolutionDigest:
      type: object
      properties:
        date:
          type: string
        resolvedToday:
          type: integer
        skipped:
          type: integer
        failed:
          type: integer
        stillOpen:
          type: integer
        approachingWindow:
          type: integer
        needsHumanReview:
          type: integer
        resolved:
          type: array
          items:
            $ref: "#/components/schemas/ResolutionDigestEntry"
        needsReviewIds:
          type: array
          items:
            type: integer
      required:
        - date
        - resolvedToday
        - skipped
        - failed
        - stillOpen
        - approachingWindow
        - needsHumanReview
        - resolved
        - needsReviewIds
    LeaderboardResponse:
      type: object
      properties:
        stats:
          $ref: "#/components/schemas/LeaderboardStats"
        calibration:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardCalibrationBucket"
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/Recommendation"
      required:
        - stats
        - calibration
        - recommendations