# lib/api-zod/src/generated/types/leaderboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alpha Lens Investment Intelligence API
 * OpenAPI spec version: 1.0.0
 */
import type { LeaderboardStatsByType } from "./leaderboardStatsByType";
import type { LeaderboardStatsPaperReturnReliability } from "./leaderboardStatsPaperReturnReliability";

export interface LeaderboardStats {
  trackRecordStart: string;
  trackRecordEnd: string;
  daysElapsed: number;
  daysRemaining: number;
  totalCalls: number;
  resolvedCalls: number;
  openCalls: number;
  correctCalls: number;
  incorrectCalls: number;
  partialCalls: number;
  winRate: number;
  winRateWithPartial: number;
  avgEdge: number;
  avgAiProbability: number;
  avgConvictionScore?: number;
  highConvictionWinRate?: number | null;
  lowConvictionWinRate?: number | null;
  totalPaperReturn: number;
  paperReturnPct: number;
  highConfidenceWinRate?: number | null;
  highEdgeWinRate?: number | null;
  paperReturnReliability: LeaderboardStatsPaperReturnReliability;
  autoResolved?: number;
  manualResolved?: number;
  pendingResolution?: number;
  byType: LeaderboardStatsByType;
}