# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alpha Lens Investment Intelligence API
 * OpenAPI spec version: 1.0.0
 */
import type { RecommendationEdgeType } from "./recommendationEdgeType";
import type { RecommendationOutcome } from "./recommendationOutcome";
import type { RecommendationResolutionMethod } from "./recommendationResolutionMethod";
import type { RecommendationType } from "./recommendationType";
import type { RecommendationUrgency } from "./recommendationUrgency";

export interface Recommendation {
  id: number;
  briefingId?: number | null;
  type: RecommendationType;
  urgency?: RecommendationUrgency;
  title: string;
  assetId?: number | null;
  assetTitle?: string;
  assetClass?: string;
  sector?: string;
  region?: string;
  direction?: string;
  aiProbability?: number | null;
  marketPrice?: number | null;
  assetPriceAtCall?: number | null;
  edge?: number | null;
  edgeType?: RecommendationEdgeType;
  convictionScore?: number | null;
  edgeCalculatedAt?: Date | null;
  edgeAgeMinutes?: number | null;
  headline?: string;
  why?: string[];
  historicalContext?: string;
  bearCase?: string;
  entryTrigger?: string;
  confidence: number;
  window?: string;
  urgencyReason?: string;
  outcome?: RecommendationOutcome;
  resolutionDate?: Date | null;
  resolutionNote?: string | null;
  marketPriceAtResolution?: number | null;
  paperReturn?: number | null;
  resolutionMethod?: RecommendationResolutionMethod;
  createdAt?: Date | null;
}